+++ jdi-light-angular-tests/src/main/java/io/github/com/pages/AngularPage.java
@@ -7,8 +7,8 @@ import com.epam.jdi.light.elements.pageobjects.annotations.locators.UI;
 import com.epam.jdi.light.ui.html.elements.common.Button;
 import com.epam.jdi.light.ui.html.elements.common.Icon;
 import com.epam.jdi.light.ui.html.elements.complex.RadioButtons;
-import io.github.com.pages.sections.InputSection;
 import io.github.com.pages.sections.CheckboxSection;
+import io.github.com.pages.sections.InputSection;
 import io.github.com.pages.sections.SlideToggleSection;
 
 public class AngularPage extends WebPage {