+++ jdi-light-html-tests/src/test/java/io/github/epam/html/tests/elements/complex/dropdown/DropdownSelectTests.java
@@ -65,7 +65,7 @@ public class DropdownSelectTests implements TestsInit {
             dressCode.select("Unknown");
             fail("You have selected dressCode that does not exist in dropdown - something went wrong");
         } catch (Exception ex) {
-            assertThat(safeException(ex), containsString("Cannot locate option with text: Unknown"));
+            assertThat(safeException(ex), containsString("Cannot locate element with text: Unknown"));
         }
     }
     @Test