+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ## 1.3.0 (planed July 2020)
 * Support [Angular materials](https://material.angular.io/components/categories) elements
 * All tasks [angular](https://github.com/jdi-testing/jdi-light/issues?q=is%3Aissue+label%3AAngular+)
+* **driver.remote.run** boolean property can be specified in **test.properties** file. If not specified, JDI would use the original logic and switch between local and remote driver depending on other properties
 
 ## 1.2.16 (release 17.05.2020)
 * Add keyboard commands (press, command, commands, pasteText) to UIElement
+* Add option to take screenshot via Robot
 
 ## 1.2.15 (release 08.06.2020)
 * Fix list issue with multiple same values